requirements.txt
colorama
psutil
requests
groq
google-generativeai
pyyaml
prompt_toolkit
openai
cryptography